[chore]: modularize app
